+++ R/frd.plot.INTFUN.R
@@ -1,23 +1,37 @@
 #internal warbleR function, not to be called by users. Plots detected frequency range.
 
-frd.plot.INTFUN <- function(wave, min.start, max.end, wl = 512, threshold = 10, wn = "hanning", flim = c(0, 22), bp = NULL, fast.spec = FALSE, ovlp = 50, pal = reverse.gray.colors.2, widths = c(2, 1), main = NULL, af.mat = NULL) {
+frd.plot.INTFUN <- function(wave, detections, wl = 512, threshold = 10, wn = "hanning", flim = c(0, 22), bp = NULL, fast.spec = FALSE, ovlp = 50, pal = reverse.gray.colors.2, widths = c(2, 1), main = NULL, all.detec = F) {
   
   # attach freq and amplitude values
-  z <- af.mat[,1]
-  zf <- af.mat[,2]
+  z <- detections$af.mat[,1]
+  zf <- detections$af.mat[,2]
   
   #sampling rate
   f <-  wave@samp.rate
+  dur <- length(wave@left)/f
+  
+  # detection limits
+  if(!all.detec)
+  {
+    min.start <- detections$frange$low.freq
+    max.end <- detections$frange$high.freq
+  } else {
+    min.start <- detections$detections[,1]
+    max.end <- detections$detections[,2]
+  }
   
   # fix flim
   flim <- c(0, floor(f/2000))
   if(flim[2] > ceiling(f/2000) - 1) flim[2] <- ceiling(f/2000) - 1 
   
+  # set limits for color rectangles down
+  if(is.null(bp)) lims <- flim else lims <- bp
+  
     # split screen
     m <- rbind(c(0, widths[1]/sum(widths), 0, 0.93), #1
                c(widths[1]/sum(widths), 1, 0 , 0.93),
                c(0, 1,  0.93 , 1)) #3 
-    
+      
     invisible(close.screen(all.screens = TRUE))  
     split.screen(m)
     screen(1)
@@ -25,9 +39,16 @@ frd.plot.INTFUN <- function(wave, min.start, max.end, wl = 512, threshold = 10, 
     
     # create spectro
     spectro.INTFUN.2(wave = wave, f = f, flim = flim, fast.spec = fast.spec, palette = pal, ovlp = ovlp, wl = wl, grid = F, tlab = "", flab = "", wn = wn)
-    
+  
+    #add gray polygon on detected frequency bands
+    lapply(seq_len(length(min.start)), function(e)
+    {
+      rect(xleft = 0, ybottom = ifelse(is.na(min.start[e]), lims[1], min.start[e]), xright = dur, ytop = ifelse(is.na(max.end[e]), lims[2], max.end[e]), col = adjustcolor("green3", 0.1), border = adjustcolor("gray", 0.2)) 
+      
+    })
+      
     #add line highlighting freq range
-    abline(h = c(min.start, max.end), col = "#80C3FF", lty = 3, lwd = 3.3)
+      abline(h = c(min.start, max.end), col = "#80C3FF", lty = 3, lwd = 2) 
     
     # add axis labels
     mtext(side = 1, text = "Time (s)", line = 2.3)
@@ -46,7 +67,8 @@ frd.plot.INTFUN <- function(wave, min.start, max.end, wl = 512, threshold = 10, 
     # fix amplitude values to close polygon (just for ploting)
     z3 <- c(0, z, 0)
     
-    if(!is.null(bp)) zf3 <- c(bp[1], zf, bp[2]) else zf3 <- c(flim[1], zf, flim[2])
+    #addd  extremes to make polygon close fine
+    zf3 <- c(lims[1], zf, lims[2])
     
     # plot amplitude values curve
     polygon(cbind(z3, zf3), col= adjustcolor("#4D69FF", 0.9))
@@ -56,7 +78,15 @@ frd.plot.INTFUN <- function(wave, min.start, max.end, wl = 512, threshold = 10, 
     
     # add bacground color
     rect(xleft = 0, ybottom = flim[1], xright = 1, ytop = flim[2], col = adjustcolor("#4D69FF", 0.05))
+
+    #add gray polygon on detected frequency bands
+    lapply(seq_len(length(min.start)), function(e)
+    {
+      rect(xleft = 0, ybottom = ifelse(is.na(min.start[e]), lims[1], min.start[e]), xright = 1, ytop = ifelse(is.na(max.end[e]), lims[2], max.end[e]), col = adjustcolor("green3", 0.1), border = adjustcolor("gray", 0.2)) 
+      
+    })
     
+        
     # add gray boxes in filtered out freq bands
     if(!is.null(bp))
     {  rect(xleft = 0, ybottom = bp[2], xright = 1, ytop = flim[2], col = adjustcolor("gray", 0.5)) 
@@ -64,8 +94,9 @@ frd.plot.INTFUN <- function(wave, min.start, max.end, wl = 512, threshold = 10, 
     }
     
     # add line to highligh freq range
-    abline(v = threshold/100, col = adjustcolor("blue4", 0.7), lty = 3, lwd = 2.3)
-    abline(h = c(min.start, max.end), col = "#80C3FF", lty = 3, lwd = 3.3)
+    abline(v = threshold/100, col = adjustcolor("blue4", 0.7), lty = 3, lwd = 2)
+    abline(h = c(min.start, max.end), col = "#80C3FF", lty = 3, lwd = 2)
+    
     
     if(!is.null(main))
     {