+++ NEWS.md
@@ -5,8 +5,9 @@
 
 ## Changes and additions:
 
-* "trackfreqs", "dfts", "ffts", "dfDTW" "ffDTW" and "specan" functions can use "frange.detec" internally to set bandpass limits 
+* "trackfreqs", "dfts", "ffts", "dfDTW" "ffDTW" and "specan" functions can use "frange.detec" internally to set bandpass limits (see "frange.detec" argument)
 * New argument "index" in  "seltailor" function to indicate which selection should be tailored
+* New arguments "threshold.freq", "threshold.time" and "wl.freq" in "specan" function to set the amplitude threshold and windows length independently for time and frequency domain measruements
 
 ## New functions: